requirements.txt
Django==3.0.7
django-filter==2.2.0
django-storages==1.9.1
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.1.2
django-heroku==0.3.1
dj-database-url==0.5.0
psycopg2==2.8.4
psycopg2-binary==2.8.5
boto3==1.12.39
beautifulsoup4 == 4.9.1
spacy==2.2.4
requests-html == 0.10.0
pytube3 == 9.6.4 
